♻ refactor: 계산로직 분리
